Update personal resume website
The website technology is out-dated. Using PHP and not using Vue.js nor
React.JS. In the future, this website, while visually looking good, for
under-the-hood reasons, should be replaced soon. However, while the site
does use PHP, there is elegancy in simplicity. For now, an update to
the existing site is good enough.