Manual chunks to optimize build
`@fortawesome~free-solid-svg-icons` is a large package. Due to laziness,
the app loads the entire library of SVG icons provided by
FontAwesome Free. This is the reason it is 904.84kB (gzip: 301.29 kB) by
itself. Other than the enormous FontAwesome library, the manual chunking
appears to have worked well.

The minifcation method is from:
https://github.com/vitejs/vite/discussions/3894#discussioncomment-1656095